// src/types/auth.ts
export type SignUpRequest = {
  displayName: string;
  email: string;
  password: string;
};

export type FirebaseAuthError = Omit<AuthError, 'code'> & {
  code: AuthErrorCodes;
};

export type AuthErrorResponse = {
  code: AuthErrorCodes;
  name: string;
  message: string;
};

export type AuthResponse<T> = ApiResponse<T, AuthErrorResponse>;